+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Changed the beta feature for using the hub's gyro. Gyro control can now be
   toggled using `use_gyro` instead of using a separate `GyroDriveBase` class.
 
+## Added
+- Added `set` to `ubuiltins` module.
+
 ## 3.3.0b5 - 2023-05-16
 
 ### Added